// src/util/interface.ts
export interface IPerson {
    name: string;
    lastName: string;
    country: string;
    birth: string;
    email: string;
    id: string
  }

  export interface IPersonService {
    name: string;
    lastName: string;
    country: string;
    birth: string;
    email: string;
    id: string;
    onDelete: (id: string) => void ;
    onEdit: (person: IPerson) => void;
  }